+++ blog-time.php
@@ -275,15 +275,22 @@ class c2c_BlogTime {
 	/**
 	 * Determines the time format string for the given context.
 	 *
-	 * Recognized contexts:
+	 * Recognized contexts (though not all have special handling):
+	 * - ajax:     When time is being returned via AJAX. Functionally equivalent
+	 *             to 'default'.
 	 * - default:  The default context. Returns the true blog time format with
 	 *             everything (default value, parameter, setting, filter) taken
 	 *             into account.
+	 * - momentjs: When time is to be ultimately used by MomentJS. Functionally
+	 *             equivalent to 'default', but with format tokens converted from
+	 *             PHP-syntax to MomentJS-syntax.
 	 * - nofilter: Takes default value, parameter, and setting into account, but
 	 *             does not pass the blog time format through any filters.
 	 * - raw:      Takes parameter and setting into account, but ignores default
 	 *             value and filter. If parameter is set, then returns that value.
 	 *             Else returns value of setting, even if empty.
+	 * - template_tag: When time is being retrieved or displayed via the
+	 *             template tag. Functionally equivalent to 'default'.
 	 * - widget:   When time is being displayed within a widget. Functionally
 	 *             equivalent to 'default'.
 	 *