+++ CHANGELOG.md
@@ -21,6 +21,7 @@
         * Change: Move `tests/*.php` into `tests/phpunit/tests/`
         * Change: Remove 'test-' prefix from unit test files
         * Change: Rename `phpunit.xml` to `phpunit.xml.dist` per best practices
+* New: Add a few more possible TODO items
 
 ## 3.6.2 _(2020-06-11)_