+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Change: Use default time format if parameter or filter attempts to configure an empty string or non-string value
 * Fix: Add support for the 'T' timezone format character to the dynamic clock (support for which was removed from Moment.js awhile ago)
 * Change: Switch to use of `wp_add_inline_script()` instead of `wp_localize_script()`
+* Change: Add optional `$exit` arg to `report_time()` to allow not exiting after outputting the time
 * Change: Note compatibility through WP 5.7+
 * Change: Update URLs to PHP documentation for datetime formatting
 * Change: Update copyright date (2021)